refactor(api): move password controller to src
