+++ api/tests/identity-access-management/integration/application/password/password.controller.test.js
@@ -1,6 +1,17 @@
-import { UserNotFoundError } from '../../../../../lib/domain/errors.js';
+import { PasswordResetDemandNotFoundError, UserNotFoundError } from '../../../../../lib/domain/errors.js';
 import { passwordController } from '../../../../../src/identity-access-management/application/password/password.controller.js';
-import { catchErr, databaseBuilder, expect, hFake, sinon } from '../../../../test-helper.js';
+import { identityAccessManagementRoutes } from '../../../../../src/identity-access-management/application/routes.js';
+import { usecases } from '../../../../../src/identity-access-management/domain/usecases/index.js';
+import { InvalidTemporaryKeyError } from '../../../../../src/shared/domain/errors.js';
+import {
+  catchErr,
+  databaseBuilder,
+  domainBuilder,
+  expect,
+  hFake,
+  HttpTestServer,
+  sinon,
+} from '../../../../test-helper.js';
 
 describe('Integration | Identity Access Management | Application | Controller | password', function () {
   describe('#createResetPasswordDemand', function () {
@@ -49,6 +60,74 @@ describe('Integration | Identity Access Management | Application | Controller | 
 
         // then
         expect(error).to.be.instanceOf(UserNotFoundError);
+      });
+    });
+  });
+
+  describe('#checkResetDemand', function () {
+    let routesUnderTest;
+    let httpTestServer;
+    let user;
+    const method = 'GET';
+    const url = '/api/password-reset-demands/ABCDEF123';
+    const email = 'user@example.net';
+
+    beforeEach(async function () {
+      sinon.stub(usecases, 'getUserByResetPasswordDemand');
+      user = domainBuilder.buildUser({ email });
+      routesUnderTest = identityAccessManagementRoutes[0];
+      httpTestServer = new HttpTestServer();
+      await httpTestServer.register(routesUnderTest);
+    });
+
+    context('Success cases', function () {
+      it('should return an HTTP response with status code 200', async function () {
+        // given
+        usecases.getUserByResetPasswordDemand.resolves(user);
+
+        // when
+        const response = await httpTestServer.request(method, url);
+
+        // then
+        expect(response.statusCode).to.equal(200);
+        expect(response.result.data.type).to.equal('users');
+        expect(response.result.data.id).to.equal(user.id.toString());
+        expect(response.result.data.attributes.email).to.equal(email);
+      });
+    });
+
+    context('Error cases', function () {
+      it('should respond an HTTP response with status code 401 when InvalidTemporaryKeyError', async function () {
+        // given
+        usecases.getUserByResetPasswordDemand.rejects(new InvalidTemporaryKeyError());
+
+        // when
+        const response = await httpTestServer.request(method, url);
+
+        // then
+        expect(response.statusCode).to.equal(401);
+      });
+
+      it('should respond an HTTP response with status code 404 when PasswordResetDemandNotFoundError', async function () {
+        // given
+        usecases.getUserByResetPasswordDemand.rejects(new PasswordResetDemandNotFoundError());
+
+        // when
+        const response = await httpTestServer.request(method, url);
+
+        // then
+        expect(response.statusCode).to.equal(404);
+      });
+
+      it('should respond an HTTP response with status code 404 when UserNotFoundError', async function () {
+        // given
+        usecases.getUserByResetPasswordDemand.rejects(new UserNotFoundError());
+
+        // when
+        const response = await httpTestServer.request(method, url);
+
+        // then
+        expect(response.statusCode).to.equal(404);
       });
     });
   });
