+++ api/tests/identity-access-management/unit/application/password/password.controller.test.js
@@ -3,6 +3,36 @@ import { usecases } from '../../../../../src/identity-access-management/domain/u
 import { expect, hFake, sinon } from '../../../../test-helper.js';
 
 describe('Unit | Identity Access Management | Application | Controller | password', function () {
+  describe('#checkResetDemand', function () {
+    const email = 'user@example.net';
+    const temporaryKey = 'ABCDEF123';
+
+    const request = {
+      params: { temporaryKey },
+    };
+    let dependencies;
+
+    beforeEach(function () {
+      sinon.stub(usecases, 'getUserByResetPasswordDemand');
+      const userSerializerStub = {
+        serialize: sinon.stub(),
+      };
+      dependencies = {
+        userSerializer: userSerializerStub,
+      };
+      usecases.getUserByResetPasswordDemand.resolves({ email });
+    });
+
+    it('should return serialized user', async function () {
+      // when
+      await passwordController.checkResetDemand(request, hFake, dependencies);
+
+      // then
+      expect(usecases.getUserByResetPasswordDemand).to.have.been.calledWithExactly({ temporaryKey });
+      expect(dependencies.userSerializer.serialize).to.have.been.calledWithExactly({ email });
+    });
+  });
+
   describe('#createResetPasswordDemand', function () {
     const email = 'user@example.net';
     const locale = 'fr';
